+++ app/jobs/csv_job.rb
@@ -4,7 +4,7 @@ class CsvJob < ApplicationJob
   def perform(model, ransack_params, user)
     klass = model.constantize
     relation = klass.ransack(ransack_params).result
-    file = CsvExportService.build_file(relation)
+    file = relation.export_csv.build_file
     AdminMailer.send_csv(model, ransack_params, file, user).deliver_now
     file.unlink
   end
