+++ app/models/antenne.rb
@@ -22,8 +22,6 @@
 
 class Antenne < ApplicationRecord
   include SoftDeletable
-  include CsvExport::Models::Antenne
-
   ## Associations
   #
   has_and_belongs_to_many :communes, inverse_of: :antennes
