+++ app/services/csv_export/antenne_exporter.rb
@@ -1,0 +1,18 @@
+module CsvExport
+  class AntenneExporter < BaseExporter
+    def fields
+      {
+        institution: -> { institution.name },
+        name: :name,
+        insee_codes: :insee_codes
+      }
+    end
+
+    def preloaded_associations
+      [
+        :institution,
+        :communes,
+      ]
+    end
+  end
+end
